Refactored BattleStream into Internal.StreamConnection
