+++ BattleshipProtocol/BattleshipProtocol.csproj
@@ -54,7 +54,7 @@
     <Compile Include="Game\Ship.cs" />
     <Compile Include="Game\ShipType.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Protocol\BattleStream.cs" />
+    <Compile Include="Protocol\Internal\StreamConnection.cs" />
     <Compile Include="Game\Commands\QuitCommand.cs" />
     <Compile Include="Protocol\Exceptions\ProtocolUnexpectedDisconnect.cs" />
     <Compile Include="Protocol\ICommandTemplate.cs" />
